Separating into two files, lots of fixes
