+++ .pre-commit-config.yaml
@@ -30,6 +30,6 @@ repos:
     - id: check-executables-have-shebangs
     - id: end-of-file-fixer
   - repo: https://github.com/psf/black
-    rev: 22.3.0
+    rev: 23.9.1
     hooks:
       - id: black
